Normalisation refactoring
