+++ src/main/java/dk/kb/netarchivesuite/solrwayback/normalise/Normalisation.java
@@ -1,13 +1,15 @@
 package dk.kb.netarchivesuite.solrwayback.normalise;
 
 
-
 import dk.kb.netarchivesuite.solrwayback.properties.PropertiesLoader;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 /**
  * 
- * This class will delegate to the Normalisation class defined in solrWayback.properties
+ * This class will delegate to the Normalisation class defined in solrWayback.properties 
+ * Always use NORMAL unless you using very old versions of WARC-Indexer
+ * 
+ * This class needs some refactoring into an abstract class instead of the switches!
  * 
  * @author teg
  *
@@ -15,17 +17,18 @@ import org.slf4j.LoggerFactory;
 public class Normalisation {
  
    private static final Logger log = LoggerFactory.getLogger(Normalisation.class);
-   private enum NormaliseType {NORMAL,LEGACY,HERITRIX};
+   private enum NormaliseType {NORMAL,LEGACY,MINIMAL};
    
    static private NormaliseType type = NormaliseType.NORMAL;
    
     static {
-       String normaliseProperty=PropertiesLoader.URL_NORMALISE;
+       String normaliseProperty=PropertiesLoader.URL_NORMALISER;
+              
        if ("legacy".equalsIgnoreCase(normaliseProperty)){
            type=NormaliseType.LEGACY;
        }
-       else if("heritrix".equalsIgnoreCase(normaliseProperty)){
-           type=NormaliseType.HERITRIX;
+       else if("minimal".equalsIgnoreCase(normaliseProperty)){
+           type=NormaliseType.MINIMAL;
        } 
        else {
            type = NormaliseType.NORMAL;           
@@ -43,8 +46,8 @@ public class Normalisation {
         case LEGACY:
           return NormalisationLegacy.canonicaliseURL(url, true, true);            
                
-         case HERITRIX:
-           return NormalisationLegacy.canonicaliseURL(url, true, true);                        
+         case MINIMAL:
+           return NormalisationMinimal.canonicaliseURL(url, true, true);                        
         }
         
         return NormalisationStandard.canonicaliseURL(url, true, true);
@@ -52,10 +55,33 @@ public class Normalisation {
     }
 
     public static String canonicaliseURL(String url, boolean allowHighOrder, boolean createUnambiguous) {
-        return NormalisationStandard.canonicaliseURL(url, allowHighOrder, createUnambiguous);               
-    }
+        switch (type) {
+        case NORMAL:
+            return NormalisationStandard.canonicaliseURL(url, true, true);
+        
+        case LEGACY:
+          return NormalisationLegacy.canonicaliseURL(url, true, true);            
+               
+         case MINIMAL:
+           return NormalisationMinimal.canonicaliseURL(url, allowHighOrder, createUnambiguous);                         
+        }
+        
+        return NormalisationStandard.canonicaliseURL(url, allowHighOrder, createUnambiguous); 
+     }
+                    
+    
 
     public static String resolveRelative(String url, String relative, boolean normalise) throws IllegalArgumentException {        
+        switch (type) {
+        case NORMAL:
+            return NormalisationStandard.resolveRelative(url, relative, normalise);
+        
+        case LEGACY:
+          return NormalisationLegacy.resolveRelative(url, relative, normalise);            
+               
+         case MINIMAL:
+           return NormalisationMinimal.resolveRelative(url, relative, normalise);                         
+        }
         return NormalisationStandard.resolveRelative(url, relative, normalise);
     }
 }
