+++ src/test/resources/properties/solrwayback.properties
@@ -41,14 +41,11 @@ screenshot.temp.imagedir=/tmp/
 #Timeout in seconds. Optional, 10 seconds is default. 
 screenshot.preview.timeout=10
 
-#The possible values for url.normaliser are: normal, legacy and heritrix. Only change if you have an very old index already build.
+#The possible values for url.normaliser are: normal, legacy and minimal. Only change if you have an very old index already build.
 # Use legacy if using warc-indexer 3.0 or 3.1
-# Use heritrix is using warc-indexer before 3.0 (Playback quality is drastically reduced) 
+# Use minimal if using warc-indexer before 3.0 (Playback quality is drastically reduced). 
+# Minimal also require the following solr params added: solr.search.params: f.url_norm.qf=url 
 url.normaliser=normal
-
-#Only set this property if using  of warc-indexer version before 3.0 that did not have url_norm field.
-#If enabled also add a solr.search.params: f.url_norm.qf=url
-#normalise.urls=false;
 
 # Optional list of Solr-params. Format is key1=value1;key2=value2,...
 #solr.search.params=f.url_norm.qf=url
